requirements.txt
beautifulsoup4==4.4.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauthlib==0.7.2
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
Werkzeug==0.10.4
